remove aggregated journal functionality from production code
The code is still necessary in a migration for the time being
