+++ modules/bim/lib/open_project/bim/engine.rb
@@ -103,7 +103,6 @@ module OpenProject::Bim
 
     patch_with_namespace :OpenProject, :CustomStyles, :ColorThemes
     patch_with_namespace :API, :V3, :Activities, :ActivityRepresenter
-    patch_with_namespace :Journal, :AggregatedJournal
     patch_with_namespace :API, :V3, :WorkPackages, :EagerLoading, :Checksum
 
     patch_with_namespace :DemoData, :QueryBuilder
